requirements.txt
PyYAML~=6.0.2
fastapi~=0.115.12
Hypercorn~=0.17.3